+++ lib/essentials/onboarding/application/message_lens_installation_evidence_reader.dart
@@ -2,7 +2,7 @@ import '../domain/message_lens_installation_state.dart';
 import '../domain/onboarding_operation_snapshot.dart';
 
 abstract interface class MessageLensInstallationEvidenceReader {
-  MessageLensInstallationEvidence read({
+  Future<MessageLensInstallationEvidence> read({
     required String archiveRootPath,
     required OnboardingOperationSnapshot operationSnapshot,
   });